renamed new parser and schema as updateParser.js and UpdateSchema.js
